Split ibc.go
